refactor: Migrate to WXT
