+++ entrypoints/popup/index.html
@@ -192,7 +192,7 @@
                 <div class="danmaku-list" id="danmaku-list"></div>
             </div>
         </div>
-        <script src="../utils/channelAssociation.js"></script>
-        <script src="popup.js"></script>
+        <script type="module" src="../../utils/channelAssociation.js"></script>
+        <script type="module" src="./popup.js"></script>
     </body>
 </html>
