+++ entrypoints/popup/popup.js
@@ -1,10 +1,9 @@
 // 引入频道关联管理工具
-// 由于在扩展环境中，通过script标签加载
-// 工具类会在页面加载时自动初始化为全局变量
+import { channelAssociation, ChannelAssociationManager } from '../../utils/channelAssociation.js';
 
 // 获取当前标签页信息
 async function getCurrentTab() {
-    const [tab] = await chrome.tabs.query({ active: true, currentWindow: true });
+    const [tab] = await browser.tabs.query({ active: true, currentWindow: true });
     return tab;
 }
 
@@ -97,7 +96,7 @@ function renderSocialIcons(config) {
             // 添加点击事件
             iconElement.addEventListener('click', () => {
                 if (link.url) {
-                    chrome.tabs.create({ url: link.url });
+                    browser.tabs.create({ url: link.url });
                 }
             });
 
@@ -183,12 +182,12 @@ async function saveSettings() {
         weightThreshold: parseInt(document.getElementById('weight-threshold').value)
     };
 
-    await chrome.storage.local.set({ danmakuSettings: settings });
+    await browser.storage.local.set({ danmakuSettings: settings });
 
     // 通知content script更新设置
     const tab = await getCurrentTab();
     if (tab && tab.url.includes('youtube.com')) {
-        chrome.tabs.sendMessage(tab.id, {
+        browser.tabs.sendMessage(tab.id, {
             type: 'updateSettings',
             settings: settings
         });
@@ -197,7 +196,7 @@ async function saveSettings() {
 
 // 加载设置
 async function loadSettings() {
-    const result = await chrome.storage.local.get('danmakuSettings');
+    const result = await browser.storage.local.get('danmakuSettings');
     const settings = result.danmakuSettings || {
         enabled: true,
         timeOffset: 0,
@@ -297,7 +296,7 @@ async function downloadDanmaku() {
 
     try {
         // 发送消息给background script下载弹幕
-        const response = await chrome.runtime.sendMessage({
+        const response = await browser.runtime.sendMessage({
             type: 'downloadDanmaku',
             bvid: bvid,
             youtubeVideoId: youtubeVideoId
@@ -311,7 +310,7 @@ async function downloadDanmaku() {
             await checkCurrentPageDanmaku();
 
             // 通知content script加载弹幕
-            chrome.tabs.sendMessage(tab.id, {
+            browser.tabs.sendMessage(tab.id, {
                 type: 'loadDanmaku',
                 youtubeVideoId: youtubeVideoId
             });
@@ -340,7 +339,7 @@ async function checkCurrentPageDanmaku() {
     }
 
     // 检查是否已有弹幕数据
-    const result = await chrome.storage.local.get(youtubeVideoId);
+    const result = await browser.storage.local.get(youtubeVideoId);
     if (result[youtubeVideoId] && result[youtubeVideoId].danmakus) {
         const data = result[youtubeVideoId];
         document.getElementById('bilibili-url').value = data.bilibili_url || '';
@@ -349,7 +348,7 @@ async function checkCurrentPageDanmaku() {
         updateManualInputUI(true, data.bilibili_url);
 
         // 当检测到有弹幕数据时，清理可能残留的未匹配状态数据
-        await chrome.storage.local.remove(['pendingNoMatchResults', 'pendingSearchResults']);
+        await browser.storage.local.remove(['pendingNoMatchResults', 'pendingSearchResults']);
     } else {
         updateManualInputUI(false);
     }
@@ -389,7 +388,7 @@ function updateManualInputUI(hasData, bilibiliUrl = '', noMatchData = null) {
 // 打开B站视频页面
 function openBilibiliVideo(url) {
     if (url) {
-        chrome.tabs.create({ url: url });
+        browser.tabs.create({ url: url });
     }
 }
 
@@ -397,11 +396,13 @@ function openBilibiliVideo(url) {
 async function openBilibiliSpace(noMatchData) {
     try {
         // 获取频道映射信息
-        const association = await getChannelAssociation(noMatchData.channelInfo.channelId);
+        const association = await channelAssociation.getChannelAssociation(
+            noMatchData.channelInfo.channelId
+        );
 
         if (association && association.bilibiliUID) {
             const spaceUrl = `https://space.bilibili.com/${association.bilibiliUID}/video`;
-            chrome.tabs.create({ url: spaceUrl });
+            browser.tabs.create({ url: spaceUrl });
         } else {
             // 如果没有关联信息，显示提示
             showStatus('该频道尚未关联B站UP主，请先在关联区域设置', 'error');
@@ -466,7 +467,7 @@ function displayDanmakuList(danmakus) {
         const tab = await getCurrentTab();
 
         if (tab && tab.url.includes('youtube.com')) {
-            chrome.tabs.sendMessage(tab.id, {
+            browser.tabs.sendMessage(tab.id, {
                 type: 'seekToTime',
                 time: time
             });
@@ -485,7 +486,7 @@ async function getPageInfo(useCache = true) {
         // 优先从background获取缓存的准确信息
         if (useCache) {
             try {
-                const backgroundResponse = await chrome.runtime.sendMessage({
+                const backgroundResponse = await browser.runtime.sendMessage({
                     type: 'getPageInfoFromBackground'
                 });
 
@@ -510,7 +511,7 @@ async function getPageInfo(useCache = true) {
 
         // fallback：直接从content script获取
         console.log('直接从content script获取页面信息');
-        const response = await chrome.tabs.sendMessage(tab.id, {
+        const response = await browser.tabs.sendMessage(tab.id, {
             type: 'getPageInfo'
         });
 
@@ -555,7 +556,7 @@ async function smartSearchAndAssociate() {
         showStatus('正在搜索B站UP主...', 'loading');
 
         // 第一步：搜索UP主
-        const userSearchResponse = await chrome.runtime.sendMessage({
+        const userSearchResponse = await browser.runtime.sendMessage({
             type: 'searchBilibiliUser',
             keyword: channelName
         });
@@ -567,7 +568,7 @@ async function smartSearchAndAssociate() {
             // 没找到UP主，进行视频全站搜索
             showStatus('未找到对应UP主，正在搜索相关视频...', 'loading');
 
-            const videoSearchResponse = await chrome.runtime.sendMessage({
+            const videoSearchResponse = await browser.runtime.sendMessage({
                 type: 'searchBilibiliVideoGlobal',
                 keyword: videoTitle
             });
@@ -625,7 +626,7 @@ function displayUserSearchResults(users, pageInfo) {
                 bilibiliSpaceUrl: spaceUrl
             };
 
-            const success = await saveChannelAssociation(
+            const success = await channelAssociation.saveChannelAssociation(
                 pageInfo.channel.channelId,
                 associationData
             );
@@ -693,7 +694,7 @@ function displayVideoSearchResults(videos, youtubeVideoId) {
                             bilibiliSpaceUrl: `https://space.bilibili.com/${mid}`
                         };
 
-                        const success = await saveChannelAssociation(
+                        const success = await channelAssociation.saveChannelAssociation(
                             pageInfo.channel.channelId,
                             associationData
                         );
@@ -770,7 +771,7 @@ function displayChannelInfo(pageInfo) {
 // 检查关联状态
 async function checkAssociation(channelId) {
     try {
-        const association = await getChannelAssociation(channelId);
+        const association = await channelAssociation.getChannelAssociation(channelId);
 
         const statusText = document.querySelector('.status-text');
         const associationSection = document.getElementById('association-section');
@@ -857,7 +858,7 @@ async function associateUploader() {
             bilibiliSpaceUrl: spaceUrl
         };
 
-        const success = await saveChannelAssociation(channelId, associationData);
+        const success = await channelAssociation.saveChannelAssociation(channelId, associationData);
         if (!success) {
             throw new Error('保存关联信息失败');
         }
@@ -883,7 +884,7 @@ async function unassociateUploader() {
 
         const channelId = pageInfo.channel.channelId;
 
-        const success = await removeChannelAssociation(channelId);
+        const success = await channelAssociation.removeChannelAssociation(channelId);
         if (!success) {
             throw new Error('删除关联信息失败');
         }
@@ -916,7 +917,7 @@ async function autoSearchDanmaku(silent = false) {
         }
 
         // 获取关联信息
-        const association = await getChannelAssociation(channelId);
+        const association = await channelAssociation.getChannelAssociation(channelId);
 
         if (!association) {
             if (!silent) showStatus('尚未关联B站UP主', 'error');
@@ -926,7 +927,7 @@ async function autoSearchDanmaku(silent = false) {
         if (!silent) showStatus('正在搜索B站视频...', 'loading');
 
         // 发送搜索请求到background script
-        const searchResponse = await chrome.runtime.sendMessage({
+        const searchResponse = await browser.runtime.sendMessage({
             type: 'searchBilibiliVideo',
             bilibiliUID: association.bilibiliUID,
             videoTitle: videoTitle,
@@ -1010,7 +1011,7 @@ function displaySearchResults(results, youtubeVideoId) {
 // 检查待显示的搜索结果（作为备用方案）
 async function checkPendingSearchResults() {
     try {
-        const result = await chrome.storage.local.get('pendingSearchResults');
+        const result = await browser.storage.local.get('pendingSearchResults');
         const pendingResults = result.pendingSearchResults;
 
         if (pendingResults && pendingResults.results && pendingResults.results.length > 0) {
@@ -1023,7 +1024,7 @@ async function checkPendingSearchResults() {
             showStatus(`找到 ${pendingResults.results.length} 个匹配的B站视频，请选择：`, 'info');
 
             // 清理已显示的结果
-            await chrome.storage.local.remove('pendingSearchResults');
+            await browser.storage.local.remove('pendingSearchResults');
         }
     } catch (error) {
         console.error('检查待显示搜索结果失败:', error);
@@ -1033,7 +1034,7 @@ async function checkPendingSearchResults() {
 // 检查待显示的未匹配结果（作为备用方案）
 async function checkPendingNoMatchResults() {
     try {
-        const result = await chrome.storage.local.get('pendingNoMatchResults');
+        const result = await browser.storage.local.get('pendingNoMatchResults');
         const pendingNoMatchResults = result.pendingNoMatchResults;
 
         if (pendingNoMatchResults) {
@@ -1046,7 +1047,7 @@ async function checkPendingNoMatchResults() {
             showStatus('未找到匹配的B站视频，请手动输入或查看B站空间', 'info');
 
             // 清理已显示的结果
-            await chrome.storage.local.remove('pendingNoMatchResults');
+            await browser.storage.local.remove('pendingNoMatchResults');
         }
     } catch (error) {
         console.error('检查待显示未匹配结果失败:', error);
@@ -1142,7 +1143,7 @@ async function downloadBangumiDanmakuFromUI(title, episodeNumber, youtubeVideoId
         updateButton.disabled = true;
         updateButton.textContent = '更新中...';
 
-        const response = await chrome.runtime.sendMessage({
+        const response = await browser.runtime.sendMessage({
             type: 'downloadBangumiDanmaku',
             title: title,
             episodeNumber: episodeNumber,
@@ -1158,7 +1159,7 @@ async function downloadBangumiDanmakuFromUI(title, episodeNumber, youtubeVideoId
             // 通知content script加载弹幕
             const tab = await getCurrentTab();
             if (tab && tab.url.includes('youtube.com')) {
-                chrome.tabs.sendMessage(tab.id, {
+                browser.tabs.sendMessage(tab.id, {
                     type: 'loadDanmaku',
                     youtubeVideoId: youtubeVideoId
                 });
@@ -1220,7 +1221,7 @@ async function downloadDanmakuFromBV(bvid, youtubeVideoId = null) {
 
         showStatus('正在下载弹幕...', 'loading');
 
-        const response = await chrome.runtime.sendMessage({
+        const response = await browser.runtime.sendMessage({
             type: 'downloadDanmaku',
             bvid: bvid,
             youtubeVideoId: youtubeVideoId
@@ -1234,7 +1235,7 @@ async function downloadDanmakuFromBV(bvid, youtubeVideoId = null) {
             await checkCurrentPageDanmaku();
 
             // 通知content script加载弹幕
-            chrome.tabs.sendMessage(tab.id, {
+            browser.tabs.sendMessage(tab.id, {
                 type: 'loadDanmaku',
                 youtubeVideoId: youtubeVideoId
             });
@@ -1243,7 +1244,7 @@ async function downloadDanmakuFromBV(bvid, youtubeVideoId = null) {
             document.getElementById('search-results').style.display = 'none';
 
             // 清理后台的搜索结果数据
-            chrome.runtime
+            browser.runtime
                 .sendMessage({ type: 'clearSearchResults' })
                 .catch((error) => console.log('清理搜索结果失败:', error));
 
@@ -1265,7 +1266,7 @@ async function downloadDanmakuFromBV(bvid, youtubeVideoId = null) {
 }
 
 // 立即设置消息监听器，不等待DOM加载
-chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
+browser.runtime.onMessage.addListener((request, sender, sendResponse) => {
     if (request.type === 'displayMultipleResults') {
         console.log('收到搜索结果消息:', request.data.results.length);
 
@@ -1305,7 +1306,7 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 
 // 消息监听器设置完成后，立即通知background popup已准备好
 console.log('消息监听器已设置，通知background popup准备完成');
-chrome.runtime
+browser.runtime
     .sendMessage({ type: 'popupReady' })
     .then((response) => {
         if (response && response.success) {
@@ -1394,7 +1395,7 @@ function showPageInfoRefreshButton() {
 
 // 打开YouTube主页
 function openYouTube() {
-    chrome.tabs.create({ url: 'https://www.youtube.com' });
+    browser.tabs.create({ url: 'https://www.youtube.com' });
 }
 
 // 初始化
