+++ lib/protobuf-parser.js
@@ -170,3 +170,8 @@ class ProtobufParser {
 if (typeof module !== 'undefined' && module.exports) {
     module.exports = ProtobufParser;
 }
+
+// 在WXT环境中，将ProtobufParser设为全局变量
+if (typeof globalThis !== 'undefined') {
+    globalThis.ProtobufParser = ProtobufParser;
+}
