+++ utils/channelAssociation.js
@@ -70,7 +70,7 @@ class ChannelAssociationManager {
                 throw new Error('缺少必要的关联参数');
             }
 
-            const result = await chrome.storage.local.get(this.STORAGE_KEY);
+            const result = await browser.storage.local.get(this.STORAGE_KEY);
             const mappings = result[this.STORAGE_KEY] || {};
 
             mappings[channelId] = {
@@ -80,7 +80,7 @@ class ChannelAssociationManager {
                 lastUpdate: Date.now()
             };
 
-            await chrome.storage.local.set({ [this.STORAGE_KEY]: mappings });
+            await browser.storage.local.set({ [this.STORAGE_KEY]: mappings });
             return true;
         } catch (error) {
             console.error('保存频道关联失败:', error);
@@ -97,12 +97,12 @@ class ChannelAssociationManager {
         try {
             if (!channelId) return false;
 
-            const result = await chrome.storage.local.get(this.STORAGE_KEY);
+            const result = await browser.storage.local.get(this.STORAGE_KEY);
             const mappings = result[this.STORAGE_KEY] || {};
 
             delete mappings[channelId];
 
-            await chrome.storage.local.set({ [this.STORAGE_KEY]: mappings });
+            await browser.storage.local.set({ [this.STORAGE_KEY]: mappings });
             return true;
         } catch (error) {
             console.error('删除频道关联失败:', error);
@@ -116,7 +116,7 @@ class ChannelAssociationManager {
      */
     async getAllAssociations() {
         try {
-            const result = await chrome.storage.local.get(this.STORAGE_KEY);
+            const result = await browser.storage.local.get(this.STORAGE_KEY);
             return result[this.STORAGE_KEY] || {};
         } catch (error) {
             console.error('获取所有关联失败:', error);
@@ -186,7 +186,7 @@ class ChannelAssociationManager {
      */
     async getLocalAssociation(channelId) {
         try {
-            const result = await chrome.storage.local.get(this.STORAGE_KEY);
+            const result = await browser.storage.local.get(this.STORAGE_KEY);
             const mappings = result[this.STORAGE_KEY] || {};
             return mappings[channelId] || null;
         } catch (error) {
@@ -288,7 +288,6 @@ window.saveChannelAssociation = (channelId, data) =>
 window.removeChannelAssociation = (channelId) =>
     channelAssociation.removeChannelAssociation(channelId);
 
-// 如果在扩展环境中，也导出类本身
-if (typeof module !== 'undefined' && module.exports) {
-    module.exports = { ChannelAssociationManager, channelAssociation };
-}
+// ES模块导出
+export { ChannelAssociationManager, channelAssociation };
+export default channelAssociation;
