+++ wxt.config.js
@@ -1,4 +1,13 @@
 import { defineConfig } from 'wxt';
 
 // See https://wxt.dev/api/config.html
-export default defineConfig({});
+export default defineConfig({
+    manifest: {
+        permissions: ['storage', 'activeTab'],
+        host_permissions: [
+            'https://api.bilibili.com/*',
+            'https://www.bilibili.com/*',
+            'https://www.youtube.com/oembed*'
+        ]
+    }
+});
